scripts/requirements.txt
kicad-skip==0.2.5
sexpdata==1.0.2
